third_party/WebKit/LayoutTests/virtual/threaded/fast/idleToBlob/canvas-toBlob-file-vs-blob.html

<script src="../../../../resources/testharness.js"></script>
<script src="../../../../resources/testharnessreport.js"></script>

<script> 
async_test(t => {

    var resultImageObject = null;
    canvas = document.createElement('canvas');
    canvas.toBlob(function(b) {
        resultImageObject = b;
        assert_false(resultImageObject instanceof File);
        assert_true(resultImageObject instanceof Blob);
        t.done();
    });

}, "Test if toBlob creates a Blob-type object.");
</script>
